relocate some codes
