+++ src/ai/backend/manager/models/rbac_models/migration/entity/abc.py
@@ -1,9 +1,7 @@
 from abc import ABC, abstractmethod
-from collections.abc import Generator
-from typing import Generic, Self, TypeVar
+from typing import Self, TypeVar
 
 import sqlalchemy as sa
-from sqlalchemy.engine import Connection
 from sqlalchemy.engine.row import Row
 
 from ai.backend.manager.data.permission.id import ObjectId, ScopeId
@@ -47,25 +45,3 @@ class AbstractEntityType(ABC):
 
 
 TEntity = TypeVar("TEntity", bound=AbstractEntityType)
-
-
-class EntityQuerier(Generic[TEntity]):
-    def __init__(self, entity_type: type[TEntity]) -> None:
-        self._entity_type = entity_type
-
-    def query_entities(
-        self,
-        db_conn: Connection,
-    ) -> Generator[list[TEntity], None, None]:
-        offset = 0
-        limit = 100
-
-        while True:
-            stmt = self._entity_type.query_statement().offset(offset).limit(limit)
-            result = db_conn.execute(stmt)
-            rows = result.all()
-            if not rows:
-                break
-
-            offset += limit
-            yield [self._entity_type.from_row(entity_row) for entity_row in rows]
